Switch to fragmented code.
