+++ src/tab/index.ts
@@ -1,0 +1,3 @@
+export * from './model';
+export * from './tab';
+export * from './tabs';
